pyproject.toml
[tool.poetry]
name = "dj-profile"
version = "0.1.0"
description = "Profile Description"
authors = ["Safrizal"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
drf-spectacular = "^0.26.5"
psycopg2-binary = "^2.9.9"
celery = "^5.3.6"
redis = "^5.0.1"


[tool.poetry.group.test.dependencies]
pytest-mock = "^3.12.0"
faker = "^20.1.0"
pytest-django = "4.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
